+++ features/invitation/components/invitation-header.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/shadcn/button';
 import { Plus } from 'lucide-react';
 import Link from 'next/link';
